requirements.txt
numpy~=1.22.2
Wave~=0.0.2
matplotlib~=3.5.1
scipy~=1.8.0
PyYAML~=6.0
PyQt5~=5.15.6